update User, Employee, Department, add File

+++ src/main/java/com/example/user_management_system/controller/EmployeeController.java
@@ -5,15 +5,16 @@ import com.example.user_management_system.entity.Employee;
 import com.example.user_management_system.service.DepartmentService;
 import com.example.user_management_system.service.EmployeeService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Date;
-import java.util.Optional;
+import java.util.*;
+import java.util.function.Function;
 
 @Controller
 public class EmployeeController {
@@ -24,24 +25,32 @@ public class EmployeeController {
     DepartmentService departmentService;
 
     @GetMapping("/employees")
-    public String list(Model model){
-        Collection<Employee> employees = employeeService.findAll();
-        System.out.println(employees);
-        model.addAttribute("employees", employees);
+    public String list(@RequestParam(defaultValue = "0") Integer page, Model model){
+        //Collection<Employee> employees = employeeService.findAll();
+        Page<Employee> employeePages = employeeService.findAll(page, 10);
+        model.addAttribute("employeePages", employeePages);
 
         return "employees/list";
     }
 
     @GetMapping("/employees/search")
     public String search(@RequestParam Integer id, Model model){
-        Collection<Employee> employees = new ArrayList<>();
-        employees.add(employeeService.findById(id).get());
+        Collection<Department> departments = departmentService.findAll();
+        model.addAttribute("departments", departments);
 
-        System.out.println(employees);
+        Optional<Employee> employee = employeeService.findById(id);
 
-        model.addAttribute("employees", employees);
+        if(employee.isEmpty()){
+            Page<Employee> employeePages = employeeService.findAll(0, 10);
+            System.out.println(employeePages.toString());
+            model.addAttribute("employeePages", employeePages);
 
-        return "employees/list";
+            model.addAttribute("errorMessage", "该员工不存在");
+
+            return "employees/list";
+        }
+        model.addAttribute("employee", employee.get());
+        return "employees/add";
     }
 
     @GetMapping("/api/employees")
@@ -91,9 +100,19 @@ public class EmployeeController {
     }
 
     @DeleteMapping("/employee/{id}")
-    public String deleteEmployee(@PathVariable("id") Integer id){
-        System.out.println(id);
-        employeeService.deleteById(id);
-        return "redirect:/employees";
+    public String deleteEmployee(@PathVariable("id") Integer id, Model model){
+        // 解除和 Department 的绑定（department.leading_official 可能是被删除员工）
+        List<Department> departments = departmentService.findByLeadingOfficial(id);
+
+        if(departments.isEmpty()){
+            employeeService.deleteById(id);
+            return "redirect:/employees";
+        }
+
+        Page<Employee> employeePages = employeeService.findAll(0, 10);
+
+        model.addAttribute("employeePages", employeePages);
+        model.addAttribute("errorMessage", "删除失败：该员工为部门负责人，请先替换部门负责人");
+        return "employees/list";
     }
 }
